+++ @navikt/core/tokens/darkside/tokens/shadow.ts
@@ -11,7 +11,7 @@ export function shadowTokenConfig(theme: AkselColorTheme) {
             : `0px 0px 1px 0px rgba(7 9 13 / 0.00), 0px 2px 5px 0px rgba(7 9 13 / 0.35), 0px 5px 12px 0px rgba(7 9 13 / 0.50)`,
         type: "shadow",
         comment:
-          "Used for dialog elements like modals and popovers. Used in box-shadow property.",
+          "Brukes som box-shadow på eleverte elementer som modaler og popovers.",
         figmaIgnore: true,
       },
     },